+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/messaginglayer/ControlInputPort.scala
@@ -10,7 +10,7 @@ import edu.uci.ics.amber.engine.common.rpc.AsyncRPCClient.{ControlInvocation, Re
 import edu.uci.ics.amber.engine.common.rpc.{AsyncRPCClient, AsyncRPCServer}
 import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, VirtualIdentity}
 import edu.uci.ics.amber.error.WorkflowRuntimeError
-import edu.uci.ics.amber.recovery.MainLogStorage.MainLogElement
+import edu.uci.ics.amber.engine.recovery.MainLogStorage.MainLogElement
 
 import scala.collection.mutable
