+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/recovery/local/LocalDiskMainLogStorage.scala
@@ -1,8 +1,8 @@
-package edu.uci.ics.amber.recovery.local
+package edu.uci.ics.amber.engine.recovery.local
 
 import edu.uci.ics.amber.engine.architecture.messaginglayer.ControlInputPort
 import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, VirtualIdentity}
-import edu.uci.ics.amber.recovery.MainLogStorage
+import edu.uci.ics.amber.engine.recovery.MainLogStorage
 
 class LocalDiskMainLogStorage(id:ActorVirtualIdentity) extends MainLogStorage {
   override def persistentEntireMessage(message: ControlInputPort.WorkflowControlMessage): Unit = ???
