+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/recovery/local/LocalDiskSecondaryLogStorage.scala
@@ -1,7 +1,7 @@
-package edu.uci.ics.amber.recovery.local
+package edu.uci.ics.amber.engine.recovery.local
 
 import edu.uci.ics.amber.engine.common.virtualidentity.{ActorVirtualIdentity, VirtualIdentity}
-import edu.uci.ics.amber.recovery.SecondaryLogStorage
+import edu.uci.ics.amber.engine.recovery.SecondaryLogStorage
 
 class LocalDiskSecondaryLogStorage(id:ActorVirtualIdentity) extends SecondaryLogStorage {
   override def persistCurrentDataCursor(dataCursor: Long): Unit = ???
