+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/recovery/mem/InMemorySecondaryLogStorage.scala
@@ -1,7 +1,7 @@
-package edu.uci.ics.amber.recovery.mem
+package edu.uci.ics.amber.engine.recovery.mem
 
 import edu.uci.ics.amber.engine.common.virtualidentity.ActorVirtualIdentity
-import edu.uci.ics.amber.recovery.SecondaryLogStorage
+import edu.uci.ics.amber.engine.recovery.SecondaryLogStorage
 
 class InMemorySecondaryLogStorage(id:ActorVirtualIdentity) extends SecondaryLogStorage {
