+++ core/amber/src/test/scala/edu/uci/ics/amber/engine/recovery/RecoverySpec.scala
@@ -1,0 +1,30 @@
+package edu.uci.ics.amber.engine.recovery
+
+import akka.actor.{ActorSystem, Props}
+import akka.testkit.{ImplicitSender, TestKit}
+import akka.util.Timeout
+import edu.uci.ics.amber.clustering.SingleNodeListener
+import org.scalatest.BeforeAndAfterAll
+import org.scalatest.flatspec.AnyFlatSpecLike
+
+import scala.concurrent.ExecutionContextExecutor
+import scala.concurrent.duration._
+
+class RecoverySpec
+    extends TestKit(ActorSystem("RecoverySpec"))
+    with ImplicitSender
+    with AnyFlatSpecLike
+    with BeforeAndAfterAll {
+
+  implicit val timeout: Timeout = Timeout(5.seconds)
+  implicit val executionContext: ExecutionContextExecutor = system.dispatcher
+
+  override def beforeAll: Unit = {
+    system.actorOf(Props[SingleNodeListener], "cluster-info")
+  }
+  override def afterAll: Unit = {
+    TestKit.shutdownActorSystem(system)
+  }
+
+
+}
